# [2.x] Reduce number of long-living instances to speed up loading