--- a/sbt-app/src/main/scala/package.scala
+++ b/sbt-app/src/main/scala/package.scala
@@ -33,7 +33,6 @@ package object sbt
     fillTaskAxis,
     mapScope,
     transform,
-    transformRef,
     inThisBuild,
     inScope,
     normalizeModuleID